requirements.txt
tqdm
hydra-core==1.2.0
tensorboard==2.15.1
pytorch-lightning==2.1.4
pandas==2.2.1
scikit-learn==1.4.1.post1
matplotlib==3.8.4